cmd/juju/commands: Split out SSHCommon into it's own files
