+++ cmd/juju/commands/ssh_test.go
@@ -6,20 +6,15 @@ package commands
 import (
 	"bytes"
 	"fmt"
-	"os"
-	"path/filepath"
 	"reflect"
 	"strings"
 
 	"github.com/juju/cmd"
 	jc "github.com/juju/testing/checkers"
-	"github.com/juju/utils/ssh"
 	gc "gopkg.in/check.v1"
 	"gopkg.in/juju/charm.v6-unstable"
 
 	"github.com/juju/juju/apiserver"
-	"github.com/juju/juju/juju/testing"
-	"github.com/juju/juju/network"
 	"github.com/juju/juju/state"
 	"github.com/juju/juju/testcharms"
 	coretesting "github.com/juju/juju/testing"
@@ -29,30 +24,6 @@ var _ = gc.Suite(&SSHSuite{})
 
 type SSHSuite struct {
 	SSHCommonSuite
-}
-
-type SSHCommonSuite struct {
-	testing.JujuConnSuite
-	bin string
-}
-
-func (s *SSHCommonSuite) SetUpTest(c *gc.C) {
-	s.JujuConnSuite.SetUpTest(c)
-	ssh.ClearClientKeys()
-	s.PatchValue(&getJujuExecutable, func() (string, error) { return "juju", nil })
-
-	s.bin = c.MkDir()
-	s.PatchEnvPathPrepend(s.bin)
-	for _, name := range patchedCommands {
-		f, err := os.OpenFile(filepath.Join(s.bin, name), os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0777)
-		c.Assert(err, jc.ErrorIsNil)
-		_, err = f.Write([]byte(fakecommand))
-		c.Assert(err, jc.ErrorIsNil)
-		err = f.Close()
-		c.Assert(err, jc.ErrorIsNil)
-	}
-	client, _ := ssh.NewOpenSSHClient()
-	s.PatchValue(&ssh.DefaultClient, client)
 }
 
 const (
@@ -222,41 +193,4 @@ func (s *SSHSuite) testSSHCommandHostAddressRetry(c *gc.C, proxy bool) {
 	code = cmd.Main(newSSHCommand(), ctx, args)
 	c.Check(code, gc.Equals, 0)
 	c.Assert(called, gc.Equals, 2)
-}
-
-func (s *SSHCommonSuite) setAddresses(m *state.Machine, c *gc.C) {
-	addrPub := network.NewScopedAddress(
-		fmt.Sprintf("admin-%s.dns", m.Id()),
-		network.ScopePublic,
-	)
-	addrPriv := network.NewScopedAddress(
-		fmt.Sprintf("admin-%s.internal", m.Id()),
-		network.ScopeCloudLocal,
-	)
-	err := m.SetProviderAddresses(addrPub, addrPriv)
-	c.Assert(err, jc.ErrorIsNil)
-}
-
-func (s *SSHCommonSuite) makeMachines(n int, c *gc.C, setAddresses bool) []*state.Machine {
-	var machines = make([]*state.Machine, n)
-	for i := 0; i < n; i++ {
-		m, err := s.State.AddMachine("quantal", state.JobHostUnits)
-		c.Assert(err, jc.ErrorIsNil)
-		if setAddresses {
-			s.setAddresses(m, c)
-		}
-		// must set an instance id as the ssh command uses that as a signal the
-		// machine has been provisioned
-		inst, md := testing.AssertStartInstance(c, s.Environ, m.Id())
-		c.Assert(m.SetProvisioned(inst.Id(), "fake_nonce", md), gc.IsNil)
-		machines[i] = m
-	}
-	return machines
-}
-
-func (s *SSHCommonSuite) addUnit(srv *state.Service, m *state.Machine, c *gc.C) {
-	u, err := srv.AddUnit()
-	c.Assert(err, jc.ErrorIsNil)
-	err = u.AssignToMachine(m)
-	c.Assert(err, jc.ErrorIsNil)
 }
